Menambahkan beberapa fitur pada user seperti leave request, dll

+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Http\Controllers\HomeController;
+use App\Http\Controllers\LeaveRequestController;
 use App\Http\Controllers\ProfileController;
 use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
@@ -33,13 +34,47 @@ Route::middleware('auth')->group(function () {
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
 });
 
-// Admin
-Route::get('/users', [UserController::class, 'dashboard'])->
-middleware(['auth','admin'])->name('user');
-Route::get('/users/{id}', [UserController::class, 'form'])->
-middleware(['auth','admin'])->name('user.edit');
-Route::post('/users/{id}', [UserController::class, 'edit']);
-Route::delete('/users/{id}', [UserController::class, 'delete'])->
-middleware(['auth','admin'])->name('user.delete');
+
+// Admin (Users Control)
+Route::get('/admin/users', [UserController::class, 'dashboard'])->
+middleware(['auth','admin'])->name('admin.user');
+Route::get('/admin/users/{id}', [UserController::class, 'form'])->
+middleware(['auth','admin'])->name('admin.user.edit');
+Route::post('/admin/users/{id}', [UserController::class, 'edit']);
+Route::delete('/admin/users/{id}', [UserController::class, 'delete'])->
+middleware(['auth','admin'])->name('admin.user.delete');
+
+
+// Admin, Karyawan, Kadepartemen, Kahrd (LeaveRequest Control)
+Route::get('/admin/leave-request', [LeaveRequestController::class, 'dashboard'])
+->middleware(['auth','admin'])->name('admin.leave-request');
+Route::delete('/admin/leave-request/{id}', [LeaveRequestController::class, 'delete'])
+->middleware(['auth','admin'])->name('admin.leave-request.delete');
+Route::get('/admin/leave-request/form/{id}', [LeaveRequestController::class, 'form'])
+->middleware(['auth','admin'])->name('admin.leave-request.edit');
+Route::post('/admin/leave-request/form/{id}', [LeaveRequestController::class, 'edit']);
+
+Route::get('/kadepartemen/leave-request', [LeaveRequestController::class, 'dashboard'])
+->middleware(['auth','kadepartemen'])->name('kadepartemen.leave-request');
+Route::get('/kadepartemen/leave-request/form/{id}', [LeaveRequestController::class, 'form'])
+->middleware(['auth','kadepartemen'])->name('kadepartemen.leave-request.edit');
+Route::post('/kadepartemen/leave-request/form/{id}', [LeaveRequestController::class, 'edit']);
+
+Route::get('/kahrd/leave-request', [LeaveRequestController::class, 'dashboard'])
+->middleware(['auth','kahrd'])->name('kahrd.leave-request');
+Route::get('/kahrd/leave-request/form/{id}', [LeaveRequestController::class, 'form'])
+->middleware(['auth','kahrd'])->name('kahrd.leave-request.edit');
+Route::post('/kahrd/leave-request/form/{id}', [LeaveRequestController::class, 'edit']);
+
+Route::get('/leave-request', [LeaveRequestController::class, 'dashboard'])
+->middleware(['auth','karyawan'])->name('leave-request');
+Route::get('/leave-request/form', [LeaveRequestController::class, 'form'])
+->middleware(['auth','karyawan'])->name('leave-request.form');
+Route::get('/leave-request/form/{id}', [LeaveRequestController::class, 'form'])
+->middleware(['auth','karyawan'])->name('leave-request.form.detail');
+Route::post('/leave-request/form', [LeaveRequestController::class, 'store'])
+->middleware(['auth','karyawan'])->name('leave-request.store');
+
+
 
 require __DIR__.'/auth.php';
